Change example app to use create-react-app
