--- examples/react/.babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": ["es2015", "react", "stage-2"],
-  "env": {
-    "development": {
-      "presets": ["react-hmre"]
-    }
-  }
-}
