--- examples/react/index.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <title>React Howler - Simple example</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-</head>
-<body>
-  <div id="app"></div>
-</body>
-</html>
