+++ examples/react/package.json
@@ -1,30 +1,20 @@
 {
   "name": "react-howler-example",
   "private": true,
-  "scripts": {
-    "start": "webpack-dev-server --config webpack.config.js --watch --progress --color --hot --port 3000",
-    "build": "NODE_ENV=production webpack --config webpack.config.prod.js"
-  },
   "author": "Khoa Nguyen",
   "license": "MIT",
-  "dependencies": {
-    "raf": "^3.1.0",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7"
-  },
   "devDependencies": {
-    "babel-cli": "^6.16.0",
-    "babel-core": "^6.17.0",
-    "babel-loader": "^6.2.5",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-2": "^6.17.0",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^1.7.0",
-    "standard-loader": "^5.0.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.12.10",
-    "webpack-dev-server": "^1.14.0"
+    "react-scripts": "0.7.0"
+  },
+  "dependencies": {
+    "raf": "^3.3.0",
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
   }
 }
