+++ examples/react/src/App.js
