+++ examples/react/src/index.js
