+++ examples/react/src/players/index.js
