+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "prebabelify": "rimraf lib",
     "babelify": "babel src -d lib",
-    "example:react": "npm run babelify && cd examples/react && yarn install && npm run start",
-    "example:react:build": "npm run babelify && cd examples/react && yarn install && npm run build",
+    "example:react": "npm run babelify && cd examples/react && yarn install && NODE_PATH=../../lib npm start",
+    "example:react:build": "npm run babelify && cd examples/react && yarn install && NODE_PATH=../../lib npm run build",
     "deploy": "babel-node bin/deploy-gh-pages",
     "prepublish": "npm run lint && npm run babelify",
     "lint": "standard",
